requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiosqlite==0.20.0
    # via
    #   -r requirements.in
    #   tortoise-orm
annotated-types==0.7.0
    # via pydantic
iniconfig==2.0.0
    # via pytest
iso8601==2.1.0
    # via tortoise-orm
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pydantic==2.9.2
    # via tortoise-orm
pydantic-core==2.23.4
    # via pydantic
pypika-tortoise==0.2.1
    # via tortoise-orm
pytest==8.3.3
    # via pytest-asyncio
pytest-asyncio==0.24.0
    # via -r requirements.in
pytz==2024.2
    # via tortoise-orm
tortoise-orm==0.21.7
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   pydantic
    #   pydantic-core
